% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_barriers.R
\name{find_barriers}
\alias{find_barriers}
\title{Calculate Decision Boundaries for Continuous Sweep given pdelta}
\usage{
find_barriers(pdelta, tr)
}
\arguments{
\item{pdelta}{Minimum difference between true and false positive rate.}

\item{tr}{List containing training parameters:
\itemize{
  \item mup - Mean of positive class distribution
  \item omegap - Variance of positive class distribution
  \item mun - Mean of negative class distribution  
  \item omegan - Variance of negative class distribution
  \item ntrain - Number of training instances
  \item prevtrain - Prevalence of positive class in training data
}}
}
\value{
Numeric vector of length 2 containing lower and upper decision barriers
}
\description{
Calculate Decision Boundaries for Continuous Sweep given pdelta
}
\examples{
tr <- list(mup = 1, omegap = 1, mun = 0, omegan = 0.5, ntrain = 1000, prevtrain = 0.7)
find_barriers(pdelta = 0.25, tr = tr)

}
