% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_variance.R
\name{find_variance}
\alias{find_variance}
\title{Estimate the variance of Continuous Sweep}
\usage{
find_variance(barriers, tr, te, prev = ac_pmax, var_type = "advanced")
}
\arguments{
\item{barriers}{The left and right decision boundaries}

\item{tr}{List containing training parameters:
\itemize{
  \item mup - Mean of positive class distribution
  \item omegap - Variance of positive class distribution
  \item mun - Mean of negative class distribution  
  \item omegan - Variance of negative class distribution
  \item ntrain - Number of training instances
  \item prevtrain - Prevalence of positive class in training data
}}

\item{te}{Numeric vector of test scores (target population)}

\item{prev}{Initial estimate of the prevalence, preferably from adjusted count.}

\item{var_type}{The type of variance estimate:
\itemize{
  \item basic - assumes training parameters Fp and Fn as known.
  \item advanced - assumes training parameters Fp and Fn as estimated.
}}
}
\value{
A scalar containing the estimated variance
}
\description{
Estimate the variance of Continuous Sweep
}
\examples{
# Example training parameters
tr <- list(mup = 1, omegap = 1, mun = 0, omegan = 0.5, ntrain = 1000, prevtrain = 0.7)
# Generate test scores
set.seed(123)
te <- c(stats::rnorm(500, mean = tr$mup, sd = sqrt(tr$omegap)), 
        stats::rnorm(500, mean = tr$mun, sd = sqrt(tr$omegan)))
# Obtain decision boundaries
barriers <- c(-3, 3) # manual
barriers <- find_barriers(pdelta = 0.25, tr = tr) # using a pdelta
# Apply the function
find_variance(barriers, tr, te, prev = 0.6, var_type = "advanced")
}
