% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CS.R
\name{CS}
\alias{CS}
\title{Continuous Sweep Prevalence Estimation}
\usage{
CS(tr, te, barriers = "advanced", conf_alpha = FALSE)
}
\arguments{
\item{tr}{List containing training parameters:
\itemize{
  \item mup - Mean of positive class distribution
  \item omegap - Variance of positive class distribution
  \item mun - Mean of negative class distribution  
  \item omegan - Variance of negative class distribution
  \item ntrain - Number of training instances
  \item prevtrain - Prevalence of positive class in training data
}}

\item{te}{Numeric vector of test scores (target population)}

\item{barriers}{Two numeric values indicating the left and right point of the area
to be integrated OR a character "basic" or "advanced" to use optimized barriers.}

\item{conf_alpha}{Either FALSE or a numeric value between 0 and 1 indicating the
alpha parameter for the confidence interval.}
}
\value{
A prevalence estimate or a list containing:
\itemize{
  \item prevalence - Prevalence estimate
  \item variance - Variance estimate
  \item conf_interval - Lower and upper bounds of the confidence interval
  \item conf_alpha - Alpha parameter for the confidence interval
  \item pdelta - Optimal pdelta estimate
}
}
\description{
Estimates prevalence values using continuous sweep based on:
\itemize{
\item Training parameters from positive and negative distributions
\item Test scores from target population
\item Decision boundaries (barriers) for the area to integrate
}
}
\examples{
# Example training parameters
tr <- list(mup = 1, omegap = 1, mun = 0, omegan = 0.5, ntrain = 1000, prevtrain = 0.7)
# Generate test scores
set.seed(123)
te <- c(rnorm(500, mean = tr$mup, sd = sqrt(tr$omegap)), 
        rnorm(500, mean = tr$mun, sd = sqrt(tr$omegan)))

# Calculate prevalence estimate
CS(tr, te, barriers = c(-1, 1))
CS(tr, te, barriers = "advanced", conf_int = 0.95)

}
\references{
Method implemented in Advanced Variance Estimation for Continuous Sweep.
}
